worker/requirements.txt
openai==1.12.0
celery==5.5.0
redis==5.2.1
python-dotenv
omegaconf